proof-manager: Refactor existing impl into `NativeProofManager`
