+++ workers/proof-manager/src/implementations/mod.rs
@@ -1,0 +1,3 @@
+//! Implementations of the proof manager
+
+pub(crate) mod native_proof_manager;
