+++ workers/proof-manager/src/implementations/native_proof_manager.rs
@@ -1,0 +1,361 @@
+//! A prover implementation which uses the native prover service
+
+use std::sync::Arc;
+
+use circuit_types::traits::{SingleProverCircuit, setup_preprocessed_keys};
+use circuits::{
+    singleprover_prove_with_hint,
+    zk_circuits::{
+        valid_commitments::{
+            SizedValidCommitments, SizedValidCommitmentsWitness, ValidCommitmentsStatement,
+        },
+        valid_fee_redemption::{
+            SizedValidFeeRedemption, SizedValidFeeRedemptionStatement,
+            SizedValidFeeRedemptionWitness,
+        },
+        valid_malleable_match_settle_atomic::{
+            SizedValidMalleableMatchSettleAtomic, SizedValidMalleableMatchSettleAtomicStatement,
+            SizedValidMalleableMatchSettleAtomicWitness,
+        },
+        valid_match_settle::{
+            SizedValidMatchSettle, SizedValidMatchSettleStatement, SizedValidMatchSettleWitness,
+        },
+        valid_match_settle_atomic::{
+            SizedValidMatchSettleAtomic, SizedValidMatchSettleAtomicStatement,
+            SizedValidMatchSettleAtomicWitness,
+        },
+        valid_offline_fee_settlement::{
+            SizedValidOfflineFeeSettlement, SizedValidOfflineFeeSettlementStatement,
+            SizedValidOfflineFeeSettlementWitness,
+        },
+        valid_reblind::{SizedValidReblind, SizedValidReblindWitness, ValidReblindStatement},
+        valid_relayer_fee_settlement::{
+            SizedValidRelayerFeeSettlement, SizedValidRelayerFeeSettlementStatement,
+            SizedValidRelayerFeeSettlementWitness,
+        },
+        valid_wallet_create::{
+            SizedValidWalletCreate, SizedValidWalletCreateStatement, SizedValidWalletCreateWitness,
+        },
+        valid_wallet_update::{
+            SizedValidWalletUpdate, SizedValidWalletUpdateStatement, SizedValidWalletUpdateWitness,
+        },
+    },
+};
+use common::types::{CancelChannel, proof_bundles::ProofBundle};
+use constants::in_bootstrap_mode;
+use job_types::proof_manager::{ProofJob, ProofManagerJob, ProofManagerReceiver};
+use rayon::{ThreadPool, ThreadPoolBuilder};
+use tracing::{error, info, info_span, instrument};
+use util::{channels::TracedMessage, concurrency::runtime::sleep_forever_blocking, err_str};
+
+use crate::{error::ProofManagerError, worker::ProofManagerConfig};
+
+/// The name prefix for worker threads
+const WORKER_THREAD_PREFIX: &str = "proof-generation-worker";
+/// The stack size for worker threads
+const WORKER_STACK_SIZE: usize = 10 * 1024 * 1024; // 10 MB
+
+/// Error message when sending a proof response fails
+const ERR_SENDING_RESPONSE: &str = "error sending proof response, channel closed";
+
+/// A native prover, generates all proofs locally
+pub struct NativeProofManager {
+    /// The job queue on which to receive proof generation jobs
+    job_queue: ProofManagerReceiver,
+    /// The threadpool of workers generating proofs for the system
+    thread_pool: Arc<ThreadPool>,
+    /// The channel on which a coordinator may cancel execution
+    cancel_channel: CancelChannel,
+}
+
+impl NativeProofManager {
+    /// Create a new native proof manager
+    pub fn new(config: ProofManagerConfig) -> Result<Self, ProofManagerError> {
+        // Build a thread pool for the worker
+        let thread_pool = ThreadPoolBuilder::new()
+            .thread_name(|i| format!("{}-{}", WORKER_THREAD_PREFIX, i))
+            .stack_size(WORKER_STACK_SIZE)
+            .build()
+            .map_err(|err| ProofManagerError::Setup(err.to_string()))?;
+
+        Ok(Self {
+            job_queue: config.job_queue,
+            thread_pool: Arc::new(thread_pool),
+            cancel_channel: config.cancel_channel,
+        })
+    }
+
+    /// Run the proof manager's execution loop
+    pub fn run(self) -> Result<(), ProofManagerError> {
+        // If the relayer is in bootstrap mode, sleep forever
+        if in_bootstrap_mode() {
+            sleep_forever_blocking();
+        }
+
+        // Preprocess the circuits
+        self.thread_pool.spawn_fifo(|| {
+            Self::preprocess_circuits().unwrap();
+        });
+
+        loop {
+            // Check the cancel channel before blocking on a job
+            if self
+                .cancel_channel
+                .has_changed()
+                .map_err(|err| ProofManagerError::RecvError(err.to_string()))?
+            {
+                info!("Proof manager cancelled, shutting down...");
+                return Err(ProofManagerError::Cancelled("received cancel signal".to_string()));
+            }
+
+            // Dequeue the next job and hand it to the thread pool
+            let job = self
+                .job_queue
+                .recv()
+                .map_err(|err| ProofManagerError::JobQueueClosed(err.to_string()))?;
+
+            self.thread_pool.spawn_fifo(move || {
+                let _span = info_span!("handle_proof_job").entered();
+                if let Err(e) = Self::handle_proof_job(job) {
+                    error!("Error handling proof manager job: {}", e)
+                }
+            });
+        }
+    }
+
+    /// The main job handler, run by a thread in the pool
+    #[instrument(name = "handle_proof_job", skip(job))]
+    fn handle_proof_job(job: TracedMessage<ProofManagerJob>) -> Result<(), ProofManagerError> {
+        let ProofManagerJob { type_, response_channel } = job.consume();
+        let proof_bundle = match type_ {
+            ProofJob::ValidWalletCreate { witness, statement } => {
+                // Prove `VALID WALLET CREATE`
+                Self::prove_valid_wallet_create(witness, statement)
+            },
+
+            ProofJob::ValidReblind { witness, statement } => {
+                // Prove `VALID REBLIND`
+                Self::prove_valid_reblind(witness, statement)
+            },
+
+            ProofJob::ValidCommitments { witness, statement } => {
+                // Prove `VALID COMMITMENTS`
+                Self::prove_valid_commitments(witness, statement)
+            },
+
+            ProofJob::ValidWalletUpdate { witness, statement } => {
+                Self::prove_valid_wallet_update(witness, statement)
+            },
+
+            ProofJob::ValidMatchSettleSingleprover { witness, statement } => {
+                // Prove `VALID MATCH MPC`
+                Self::prove_valid_match_mpc(witness, statement)
+            },
+
+            ProofJob::ValidMatchSettleAtomic { witness, statement } => {
+                // Prove `VALID MATCH SETTLE ATOMIC`
+                Self::prove_valid_match_settle_atomic(witness, statement)
+            },
+
+            ProofJob::ValidMalleableMatchSettleAtomic { witness, statement } => {
+                // Prove `VALID MALLEABLE MATCH SETTLE ATOMIC`
+                Self::prove_valid_malleable_match_settle_atomic(witness, statement)
+            },
+
+            ProofJob::ValidRelayerFeeSettlement { witness, statement } => {
+                // Prove `VALID RELAYER FEE SETTLEMENT`
+                Self::prove_valid_relayer_fee_settlement(witness, statement)
+            },
+
+            ProofJob::ValidOfflineFeeSettlement { witness, statement } => {
+                // Prove `VALID OFFLINE FEE SETTLEMENT`
+                Self::prove_valid_offline_fee_settlement(witness, statement)
+            },
+
+            ProofJob::ValidFeeRedemption { witness, statement } => {
+                // Prove `VALID FEE REDEMPTION`
+                Self::prove_valid_fee_redemption(witness, statement)
+            },
+        }?;
+
+        response_channel
+            .send(proof_bundle)
+            .map_err(|_| ProofManagerError::Response(ERR_SENDING_RESPONSE.to_string()))
+    }
+
+    /// Initialize the proving key/verification key & circuit layout caches for
+    /// all of the circuits
+    pub(crate) fn preprocess_circuits() -> Result<(), ProofManagerError> {
+        // Set up the proving & verification keys for all of the circuits
+        setup_preprocessed_keys::<SizedValidWalletCreate>();
+        setup_preprocessed_keys::<SizedValidWalletUpdate>();
+        setup_preprocessed_keys::<SizedValidCommitments>();
+        setup_preprocessed_keys::<SizedValidReblind>();
+        setup_preprocessed_keys::<SizedValidMatchSettle>();
+        setup_preprocessed_keys::<SizedValidRelayerFeeSettlement>();
+        setup_preprocessed_keys::<SizedValidOfflineFeeSettlement>();
+        setup_preprocessed_keys::<SizedValidFeeRedemption>();
+
+        // Set up layouts for all of the circuits
+        SizedValidWalletCreate::get_circuit_layout().map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidWalletUpdate::get_circuit_layout().map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidCommitments::get_circuit_layout().map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidReblind::get_circuit_layout().map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidMatchSettle::get_circuit_layout().map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidRelayerFeeSettlement::get_circuit_layout()
+            .map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidOfflineFeeSettlement::get_circuit_layout()
+            .map_err(err_str!(ProofManagerError::Setup))?;
+        SizedValidFeeRedemption::get_circuit_layout()
+            .map_err(err_str!(ProofManagerError::Setup))?;
+
+        Ok(())
+    }
+
+    /// Create a proof of `VALID WALLET CREATE`
+    #[instrument(skip_all, err)]
+    fn prove_valid_wallet_create(
+        witness: SizedValidWalletCreateWitness,
+        statement: SizedValidWalletCreateStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID WALLET CREATE`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidWalletCreate>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_wallet_create(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID REBLIND`
+    #[instrument(skip_all, err)]
+    fn prove_valid_reblind(
+        witness: SizedValidReblindWitness,
+        statement: ValidReblindStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID REBLIND`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidReblind>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_reblind(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID COMMITMENTS`
+    #[instrument(skip_all, err)]
+    fn prove_valid_commitments(
+        witness: SizedValidCommitmentsWitness,
+        statement: ValidCommitmentsStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID COMMITMENTS`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidCommitments>(witness, statement)
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_commitments(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID WALLET UPDATE`
+    #[instrument(skip_all, err)]
+    fn prove_valid_wallet_update(
+        witness: SizedValidWalletUpdateWitness,
+        statement: SizedValidWalletUpdateStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID WALLET UPDATE`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidWalletUpdate>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_wallet_update(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID MATCH SETTLE`
+    #[instrument(skip_all, err)]
+    fn prove_valid_match_mpc(
+        witness: SizedValidMatchSettleWitness,
+        statement: SizedValidMatchSettleStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID MATCH SETTLE`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidMatchSettle>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_match_settle(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID MATCH SETTLE ATOMIC`
+    #[instrument(skip_all, err)]
+    fn prove_valid_match_settle_atomic(
+        witness: SizedValidMatchSettleAtomicWitness,
+        statement: SizedValidMatchSettleAtomicStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID MATCH SETTLE ATOMIC`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidMatchSettleAtomic>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_match_settle_atomic(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID MALLEABLE MATCH SETTLE ATOMIC`
+    #[instrument(skip_all, err)]
+    fn prove_valid_malleable_match_settle_atomic(
+        witness: SizedValidMalleableMatchSettleAtomicWitness,
+        statement: SizedValidMalleableMatchSettleAtomicStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID MALLEABLE MATCH SETTLE ATOMIC`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidMalleableMatchSettleAtomic>(
+                witness,
+                statement.clone(),
+            )
+            .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_malleable_match_settle_atomic(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID RELAYER FEE SETTLEMENT`
+    #[instrument(skip_all, err)]
+    fn prove_valid_relayer_fee_settlement(
+        witness: SizedValidRelayerFeeSettlementWitness,
+        statement: SizedValidRelayerFeeSettlementStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID RELAYER FEE SETTLEMENT`
+        let (proof, link_hint) = singleprover_prove_with_hint::<SizedValidRelayerFeeSettlement>(
+            witness,
+            statement.clone(),
+        )
+        .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_relayer_fee_settlement(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID OFFLINE FEE SETTLEMENT`
+    #[instrument(skip_all, err)]
+    fn prove_valid_offline_fee_settlement(
+        witness: SizedValidOfflineFeeSettlementWitness,
+        statement: SizedValidOfflineFeeSettlementStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID OFFLINE FEE SETTLEMENT`
+        let (proof, link_hint) = singleprover_prove_with_hint::<SizedValidOfflineFeeSettlement>(
+            witness,
+            statement.clone(),
+        )
+        .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_offline_fee_settlement(statement, proof, link_hint))
+    }
+
+    /// Create a proof of `VALID FEE REDEMPTION`
+    #[instrument(skip_all, err)]
+    fn prove_valid_fee_redemption(
+        witness: SizedValidFeeRedemptionWitness,
+        statement: SizedValidFeeRedemptionStatement,
+    ) -> Result<ProofBundle, ProofManagerError> {
+        // Prove the statement `VALID FEE REDEMPTION`
+        let (proof, link_hint) =
+            singleprover_prove_with_hint::<SizedValidFeeRedemption>(witness, statement.clone())
+                .map_err(|err| ProofManagerError::Prover(err.to_string()))?;
+
+        Ok(ProofBundle::new_valid_fee_redemption(statement, proof, link_hint))
+    }
+}
